Improved virtual tree handling.

+++ source/no.resheim.aggregator.core/src/no/resheim/aggregator/data/IAggregatorItem.java
@@ -35,10 +35,10 @@ public interface IAggregatorItem {
 	 */
 	public abstract String getTitle();
 
-	public abstract void setOrdering(long ordering);
+	public abstract void setOrdering(int ordering);
 
 	public abstract void setRegistry(FeedCollection registry);
 
-	public abstract long getOrdering();
+	public abstract int getOrdering();
 
 }
